requirements.txt
asgiref==3.3.1
Django==3.1.7
django-bootstrap-breadcrumbs==0.9.2
django-cleanup==5.1.0
django-crispy-forms==1.11.2
django-phonenumber-field==5.0.0
phonenumberslite==8.12.20
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
